Fix test_attributes.
test_enum_attributes was testing that it could create arbitrary enum attributes. LLVM doesn't permit that in general, you can only use attribute kinds that already exist. Remove this test.

test_attributes_on_call_site_values was creating a function void(i32) but creating a call with zero arguments, which triggers an assertion failure in LLVM creating the call instruction with mismatching arguments. Create it with one argument, check that it has one argument not zero, and then the tests that attribute the first argument have an argument to apply an attribute to.