Change typesafety quite a bit.
The data members of DebugInfoBuilder are gone because they couldn't accurately keep track of the state anyways given that the type is Copy. Most of the panic! statements went away as part of that.

Mark creation and replacement of placeholders as unsafe. We can't reliably detect their misusage, but we can document it.

Change the constructor to take all the information needed to create the one CU that this DebugInfoBuilder will refer to. Unfortunate that we can't add data to an existing CU, but that's a limitation of the LLVM C API.