chore: Run `rustfmt`.
This patch is the result of running `rustfmt **/*.rs` from the root of
this project. The goal is the following: When a user (like me) edits a
file, the file is automatically formatted on save; which results in
larger diffs than needed. Let's fix the coding styles once and for all
thanks to `rustfmt` :-)!